[Model Monitoring] Adjust Grafana dashboards to use V3IO datasource instead of MLRun API (#3271)
(cherry picked from commit 8900a91952fea5644c6c608cab525a123e7c1bad)
